requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
gunicorn==21.2.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
packaging==21.3
psycopg2-binary==2.9.6
pyparsing==3.0.7
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
